+++ example/gpio-info.js
@@ -4,7 +4,7 @@ var Gpio = require('../'),
   gpio,
   gpioNo;
 
-for (gpioNo = 0; gpioNo <= 53; gpioNo += 1) {
+for (gpioNo = Gpio.MIN_GPIO; gpioNo <= Gpio.MAX_GPIO; gpioNo += 1) {
   gpio = new Gpio(gpioNo);
 
   console.log('GPIO ' + gpioNo + ':' +