+++ History.md
@@ -1,3 +1,8 @@
+Unpublished
+===========
+
+  * constants MIN_GPIO, MAX_GPIO, and MAX_USER_GPIO added
+
 0.0.5 / Nov 03 2015
 ===================